feat: share configuration and routing UI between new and edit application pages
Extracts tabbed ConfigurationCard and smart RoutingCard into shared
components used by both /applications/new and /applications/[id].

- Adds ApplicationConfigData base schema so both create and edit forms
  share ports/envVars/volumeMounts/healthCheck/restartPolicy fields
- Adds enableHealthCheck + healthCheck to the edit schema (was missing)
- Moves PasteEnvDialog to shared components/
- ConfigurationCard: tabbed UI (env vars + paste .env, volumes, health
  check, ports, advanced) — replaces three separate flat cards on edit
- RoutingCard: subdomain+zone DNS builder, Cloudflare/Azure connectivity
  warnings, showEnableToggle prop for the edit page
- Edit page refactored to sub-component pattern so useForm is
  initialised with actual data (fixes Radix Select desync after reset)
- Routing shown on edit page whenever routing is enabled, not just for
  StatelessWeb service type
- Hostname decomposition no longer corrupts existing values when the
  zone doesn't match an available DNS zone
